requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1 
attrs==23.2.0    
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
datasets==2.16.1
dill==0.3.7
evaluate==0.4.1
# faiss-gpu==1.8.0
 # Example addition, adjust version as needed
filelock==3.14.0
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.23.2
idna==3.7
multidict==6.0.5
multiprocess==0.70.15
# numpy==1.26.4
packaging==24.0
protobuf>=3.20,<5  # This allows any version from 3.20 up to (but not including) 5
pyarrow==16.1.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
responses==0.18.0
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
tokenizers==0.15.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4

streamlit==1.34.0
st-annotated-text==4.0.1
matplotlib>=3.2.2
numpy>=1.18.5
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
Pillow>=7.1.2
pandas>=1.1.4
seaborn>=0.11.0
opencv-python>=4.1.1
gdown==5.2.0
htbuilder==0.6.2
transformers==4.39.3
